sheep/http: change type of 'req->data_length' to uint64_t
Change the type of 'data_length' in 'struct http_request' so we can
receive http request larger than 4G in 32bit environment.

+++ sheep/http/http.h
@@ -47,7 +47,7 @@ struct http_request {
 	char *uri;
 	enum http_opcode opcode;
 	enum http_status status;
-	size_t data_length;
+	uint64_t data_length;
 };
 
 struct http_driver {
